rename modules to allow for further extension
see CHANGELOG.md for thoughts
