+++ install
@@ -108,23 +108,23 @@ module_dependencies() {
 	hpi_module my.window_watcher || return $?
 	hpi_module my.smscalls || return $?
 	hpi_module my.browsing || return $?
-	hpi_module my.discord || return $?
+	hpi_module my.discord.data_export || return $?
 	hpi_module my.google_takeout || return $?
 	hpi_module my.time.tz.via_location || return $?
 	hpi_module my.coding.commits || return $?
-	hpi_module my.todotxt || return $?
+	hpi_module my.todotxt.file_backups || return $?
 	hpi_module my.location.ip || return $?
 	hpi_module my.location.gpslogger || return $?
-	hpi_module my.chess || return $?
-	hpi_module my.mpv || return $?
-	hpi_module my.league || return $?
-	hpi_module my.trakt || return $?
+	hpi_module my.chess.export || return $?
+	hpi_module my.mpv.history_daemon || return $?
+	hpi_module my.league.export || return $?
+	hpi_module my.trakt.export || return $?
 	hpi_module my.imap || return $?
-	hpi_module my.grouvee || return $?
-	hpi_module my.mal || return $?
-	hpi_module my.listenbrainz || return $?
+	hpi_module my.grouvee.export || return $?
+	hpi_module my.mal.export || return $?
+	hpi_module my.listenbrainz.export || return $?
 	hpi_module my.old_forums || return $?
-	hpi_module my.skype || return $?
+	hpi_module my.skype.gdpr || return $?
 }
 
 verify_personal_python_packages() {
