+++ my/blizzard/gdpr.py
