+++ my/location/ip.py
@@ -17,14 +17,14 @@ from my.core.common import Stats, mcachew, LazyLogger
 
 from .models import Location
 
-from ..facebook import (
+from ..facebook.gdpr import (
     AdminAction,
     UploadedPhoto,
     events as facebook_events,
     config as facebook_config,
 )
-from ..blizzard import events as blizzard_events
-from ..discord import activity, _cachew_depends_on
+from ..blizzard.gdpr import events as blizzard_events
+from ..discord.data_export import activity, _cachew_depends_on
 
 
 logger = LazyLogger(__name__, level="warning")
