+++ my/skype/gdpr.py
